+++ pyccel/ast/basic.py
@@ -4,16 +4,16 @@
 #------------------------------------------------------------------------------------------#
 
 """
-This module contains classes from which all pyccel nodes inherit.
-They are:
-- Basic, which provides a python AST
-- PyccelAstNode which describes each PyccelAstNode
+This module contains classes from which all pyccel nodes inherit. They are:
+- PyccelAstNode, which provides a base class for our Python AST nodes;
+- TypedAstNode, which inherits from PyccelAstNode and provides a base class for
+  AST nodes requiring type descriptors.
 """
 import ast
 
 from pyccel.utilities.stage   import PyccelStage
 
-__all__ = ('Basic', 'Immutable', 'PyccelAstNode', 'ScopedNode')
+__all__ = ('PyccelAstNode', 'Immutable', 'TypedAstNode', 'ScopedAstNode')
 
 dict_keys   = type({}.keys())
 dict_values = type({}.values())
@@ -25,13 +25,13 @@ pyccel_stage = PyccelStage()
 #==============================================================================
 class Immutable:
     """ Superclass for classes which cannot inherit
-    from Basic """
+    from PyccelAstNode """
     __slots__ = ()
 
 #==============================================================================
-class Basic:
+class PyccelAstNode:
     """
-    Basic class from which all objects in the Pyccel AST inherit.
+    PyccelAstNode class from which all objects in the Pyccel AST inherit.
 
     This foundational class provides all the functionalities that are common to
     objects in the Pyccel AST. This includes the construction and navigation of
@@ -52,7 +52,7 @@ class Basic:
 
             from pyccel.ast.literals import convert_to_literal
 
-            if Basic._ignore(c):
+            if PyccelAstNode._ignore(c):
                 continue
 
             elif isinstance(c, (int, float, complex, str, bool)):
@@ -63,19 +63,19 @@ class Basic:
             elif iterable(c):
                 size = len(c)
                 c = tuple(ci if (not isinstance(ci, (int, float, complex, str, bool)) \
-                                 or Basic._ignore(ci)) \
+                                 or PyccelAstNode._ignore(ci)) \
                         else convert_to_literal(ci) for ci in c if not iterable(ci))
                 if len(c) != size:
-                    raise TypeError("Basic child cannot be a tuple of tuples")
+                    raise TypeError("PyccelAstNode child cannot be a tuple of tuples")
                 setattr(self, c_name, c)
 
-            elif not isinstance(c, Basic):
-                raise TypeError("Basic child must be a Basic or a tuple not {}".format(type(c)))
+            elif not isinstance(c, PyccelAstNode):
+                raise TypeError("PyccelAstNode child must be a PyccelAstNode or a tuple not {}".format(type(c)))
 
 
             if isinstance(c, tuple):
                 for ci in c:
-                    if not Basic._ignore(ci):
+                    if not PyccelAstNode._ignore(ci):
                         ci.set_current_user_node(self)
             else:
                 c.set_current_user_node(self)
@@ -184,7 +184,7 @@ class Basic:
 
         Parameters
         ----------
-        node : Basic
+        node : PyccelAstNode
                The object whose attributes we are interested in
 
         Results
@@ -199,7 +199,7 @@ class Basic:
 
         Parameters
         ----------
-        node           : Basic
+        node           : PyccelAstNode
                       The object whose users we are interested in
         excluded_nodes : tuple of types
                       Types for which is_user_of should not be called
@@ -326,10 +326,24 @@ class Basic:
 
     @property
     def fst(self):
+        """
+        Get the AST object which Python parsed in the original code.
+
+        Get the AST (abstract syntax tree) object which Python parsed
+        in the original code. This object describes the Python code being
+        translated. It provides line numbers and columns which can be
+        used to report the origin of any potential errors.
+
+        Returns
+        -------
+        ast.AST
+            The AST object which was parsed.
+        """
         if len(self._fst) == 1:
             return self._fst[0]
         else:
             return None
+
 
     def toggle_recursion(self):
         """ Change the recursion state
@@ -344,7 +358,7 @@ class Basic:
 
     def get_all_user_nodes(self):
         """ Returns all the objects user nodes.
-        This function should only be called in Basic
+        This function should only be called in PyccelAstNode
         """
         return self._user_nodes
 
@@ -391,7 +405,7 @@ class Basic:
 
         Parameters
         ----------
-        user_node : Basic
+        user_node : PyccelAstNode
                     Node which previously used the current node
         invalidate : bool
                     Indicates whether the removed object should
@@ -435,7 +449,7 @@ class Basic:
         """
         self._pyccel_staging = pyccel_stage.current_stage
 
-class PyccelAstNode(Basic):
+class TypedAstNode(PyccelAstNode):
     """Class from which all nodes containing objects inherit
     """
     __slots__  = ()
@@ -470,6 +484,18 @@ class PyccelAstNode(Basic):
         return self._order # pylint: disable=no-member
 
     def copy_attributes(self, x):
+        """
+        Copy the attributes describing a TypedAstNode into this node.
+
+        Copy the attributes which describe the TypedAstNode passed as
+        argument (dtype, precision, shape, rank, order) into this node
+        so that the two nodes can be stored in the same object.
+
+        Parameters
+        ----------
+        x : TypedAstNode
+            The node from which the attributes should be copied.
+        """
         self._shape     = x.shape
         self._rank      = x.rank
         self._dtype     = x.dtype
@@ -478,7 +504,7 @@ class PyccelAstNode(Basic):
 
 
 #------------------------------------------------------------------------------
-class ScopedNode(Basic):
+class ScopedAstNode(PyccelAstNode):
     """ Class from which all objects with a scope inherit
     """
     __slots__ = ('_scope',)